Checkout-Country-Select  -Fixed some CR comments  -TODO: need to get country list into global state, then pass it to parent of field (select)
