+++ src/app/component/Select/Select.style.scss
@@ -11,6 +11,7 @@
  .Select {
     &-Container {
         min-width: 160px;
+        max-width: 40rem;
         height: 35px;
         font-weight: 700;
         font-size: 1.05rem;
