+++ src/app/route/CheckoutPage/CheckoutPage.style.scss
@@ -142,11 +142,7 @@
         }
     }
 
-    .Select-Container {
-            max-width: 40rem;
-
-        ul {
-            overflow-y: scroll;
-        }
+    ul {
+        overflow-y: auto;
     }
 }
