+++ src/app/route/RouterTemplate/index.js
@@ -49,7 +49,7 @@ class AppRouter extends BaseRouter {
         };
     }
 }
-
+//TODO add getcountrieslist
 const mapDispatchToProps = dispatch => ({
     updateHeaderAndFooter: (options) => {
         HeaderAndFooterDispatcher.handleData(dispatch, options);
