--- src/app/store/Field/index.js
@@ -1,10 +1,0 @@
-import FieldReducer from './Field.reducer';
-import FieldDispatcher from './Field.dispatcher';
-import { getCountryList, GET_COUNTRY_LIST } from './Field.action';
-
-export {
-    FieldReducer,
-    FieldDispatcher,
-    getCountryList,
-    GET_COUNTRY_LIST
-};
