+++ src/app/store/HeaderAndFooter/HeaderAndFooter.action.js
@@ -11,6 +11,7 @@
 
 export const UPDATE_MENU = 'UPDATE_MENU';
 export const TOGGLE_HEADER_AND_FOOTER = 'TOGGLE_HEADER_AND_FOOTER';
+const GET_COUNTRY_LIST = 'GET_COUNTRY_LIST';
 
 const updateMenu = menu => ({
     type: UPDATE_MENU,
@@ -22,4 +23,14 @@ const toggleHeaderAndFooter = isHeaderAndFooterVisible => ({
     isHeaderAndFooterVisible
 });
 
-export { updateMenu, toggleHeaderAndFooter };
+const getCountryList = countries => ({
+    type: GET_COUNTRY_LIST,
+    countries
+});
+
+export {
+    updateMenu,
+    toggleHeaderAndFooter,
+    getCountryList,
+    GET_COUNTRY_LIST
+};
