+++ src/app/store/HeaderAndFooter/HeaderAndFooter.reducer.js
@@ -12,15 +12,24 @@
 /* eslint-disable no-param-reassign */
 import {
     UPDATE_MENU,
-    TOGGLE_HEADER_AND_FOOTER
+    TOGGLE_HEADER_AND_FOOTER,
+    GET_COUNTRY_LIST
 } from './HeaderAndFooter.action';
 
 const initialState = {
     menu: {},
-    isHeaderAndFooterVisible: true
+    isHeaderAndFooterVisible: true,
+    countries: []
 };
 
 const HeaderAndFooterReducer = (state = initialState, action) => {
+    const { countries } = action;
+
+    const resultingCountries = countries && countries.map(country => ({
+        id: country.id,
+        label: country.full_name_locale
+    }));
+
     switch (action.type) {
     case UPDATE_MENU:
         const { menu: { items } } = action;
@@ -71,6 +80,12 @@ const HeaderAndFooterReducer = (state = initialState, action) => {
         return {
             ...state,
             isHeaderAndFooterVisible
+        };
+
+    case GET_COUNTRY_LIST:
+        return {
+            ...state,
+            countries: resultingCountries
         };
 
     default:
