+++ src/app/store/HeaderAndFooter/index.js
@@ -16,7 +16,9 @@ import {
     UPDATE_MENU,
     TOGGLE_HEADER_AND_FOOTER,
     updateMenu,
-    toggleHeaderAndFooter
+    toggleHeaderAndFooter,
+    getCountryList,
+    GET_COUNTRY_LIST
 } from './HeaderAndFooter.action';
 
 export {
@@ -25,5 +27,7 @@ export {
     UPDATE_MENU,
     TOGGLE_HEADER_AND_FOOTER,
     updateMenu,
-    toggleHeaderAndFooter
+    toggleHeaderAndFooter,
+    getCountryList,
+    GET_COUNTRY_LIST
 };
