+++ src/app/store/index.js
@@ -25,7 +25,6 @@ import { NoMatchReducer } from 'Store/NoMatch';
 import { RelatedProductsReducer } from 'Store/RelatedProducts';
 import { SearchBarReducer } from 'Store/SearchBar';
 import { MyAccountReducer } from 'Store/MyAccount';
-import { FieldReducer } from 'Store/Field';
 
 const reducers = combineReducers({
     CmsBlocksAndSliderReducer,
@@ -39,8 +38,7 @@ const reducers = combineReducers({
     NoMatchReducer,
     RelatedProductsReducer,
     SearchBarReducer,
-    MyAccountReducer,
-    FieldReducer
+    MyAccountReducer
 });
 
 const store = createStore(
